pyproject.toml
[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"

[project]
name = "kani-vi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nemo-toolkit[tts]<=2.5.2",
    "torch",
    "torchvision",
    "torchaudio",
    "kiwisolver",
    "onnx<1.19",
    "transformers",
    "huggingface-hub>=1.1.2",
    "soundfile",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
transformers = { git = "https://github.com/huggingface/transformers.git" }
